server/pyproject.toml
[tool.poetry]
name = "alignment protoype"
version = "0.1.0"
description = ""
authors = ["Mike Brinker <themikejr@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
gunicorn = "^20.1.0"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.5"
whitenoise = "^6.4.0"
wsgi-basic-auth = "^1.1.0"
tqdm = "^4.65.0"
django-cors-headers = "^3.14.0"
djangorestframework = "^3.14.0"
strawberry-graphql-django = "^0.9.2"

[tool.poetry.dev-dependencies]
black = "^23.1.0"
django-types = "^0.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
